pyproject.toml
[project]
name = "coral-finetuned-agent"
version = "0.1.0"
description = "Agent using fine tunned LLM for better output"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai==0.126.0",
    "python-dotenv",
    "crewai-tools==0.47.1",
    "crewai-tools[mcp]", 
]